last

+++ app/Http/Controllers/StripePaymentController.php
@@ -2,8 +2,15 @@
 
 namespace App\Http\Controllers;
 
+use App\Delivery;
+use App\Order;
+use App\OrderItem;
+use App\Product;
 use Illuminate\Http\Request;
-use Session;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Session;
+use Stripe\Charge;
+use Customer;
 use Stripe;
 use Cart;
 
@@ -15,28 +22,48 @@ class StripePaymentController extends Controller
     }
 
 
-    /**
-     * success response method.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function stripePost(Request $request)
     {
-        $pay = Cart::subTotal();
+        $delivery = Delivery::findOrFail(session('delivery_id'));
+        $ship_cost = session('shipment');
+
+        //$pay = Cart::subTotal() ;
         Stripe\Stripe::setApiKey(env('STRIPE_SECRET'));
-        Stripe\Charge::create ([
-            "amount" =>100* $pay ,
+        $pay= Stripe\Charge::create ([
+            "amount" =>(Cart::subTotal() + $ship_cost) * 100,
             "currency" => "eur",
             "source" => $request->stripeToken,
             "description" => "Test payment from Anda."
         ]);
+        if($pay->paid == true) {
+            $order = Order::create([
+                'user_id' => Auth::user()->id,
+                'delivery_id' => $delivery->id,
+                'items' => Cart::count(),
+                'totalprice' =>(Cart::subTotal() + $ship_cost) ,
+            ]);
+            foreach (Cart::content() as $cart) {
+                OrderItem::create([
+                    'order_id' => $order->id,
+                    'product_id' => $cart->id,
+                    'quantity' => $cart->qty,
+                    'price' => $cart->price,
 
-        Session::flash('success', 'Payment successful!');
+                ]);
+                $product = Product::findOrFail($cart->id);
+                $product->quantity-= $cart->qty;
+                $product->save();
+            }
 
-        Cart::destroy();
 
-        return back();
-    }
+            Session::flash('success', 'Payment successful!');
 
+            Cart::destroy();
+
+            return redirect('/');
+        }
+
+
+}
 
 }
